Maintenance: CI/CD improvements, code quality, and build fixes

+++ AestraCore/include/AestraThreading.h
@@ -2,13 +2,13 @@
 #pragma once
 
 #include <atomic>
+#include <condition_variable>
+#include <functional>
+#include <memory>
+#include <mutex>
+#include <queue>
 #include <thread>
 #include <vector>
-#include <queue>
-#include <functional>
-#include <mutex>
-#include <condition_variable>
-#include <memory>
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
@@ -31,12 +31,12 @@ public:
         static auto impl = []() {
             HMODULE hAvrt = LoadLibraryA("Avrt.dll");
             if (hAvrt) {
-                typedef HANDLE (WINAPI *AvSetMmThreadCharacteristicsA_t)(LPCSTR, LPDWORD);
+                typedef HANDLE(WINAPI * AvSetMmThreadCharacteristicsA_t)(LPCSTR, LPDWORD);
                 auto pFunc = (AvSetMmThreadCharacteristicsA_t)GetProcAddress(hAvrt, "AvSetMmThreadCharacteristicsA");
                 if (pFunc) {
                     DWORD taskIndex = 0;
                     pFunc("Pro Audio", &taskIndex);
-                    // We knowingly leak the handle return/don't revert for this thread's lifetime 
+                    // We knowingly leak the handle return/don't revert for this thread's lifetime
                     // (typical for dedicated audio threads).
                     // Also leak lib handle to keep function pointer valid.
                 }
@@ -44,7 +44,7 @@ public:
             return 0;
         }();
         (void)impl;
-        
+
         // Also boost Win32 priority
         SetThreadPriority(GetCurrentThread(), THREAD_PRIORITY_TIME_CRITICAL);
 #endif
@@ -54,8 +54,7 @@ public:
 // =============================================================================
 // Lock-Free Ring Buffer (Single Producer, Single Consumer)
 // =============================================================================
-template<typename T, size_t Size>
-class LockFreeRingBuffer {
+template <typename T, size_t Size> class LockFreeRingBuffer {
 public:
     LockFreeRingBuffer() : writeIndex(0), readIndex(0) {}
 
@@ -63,11 +62,11 @@ public:
     bool push(const T& item) {
         size_t currentWrite = writeIndex.load(std::memory_order_relaxed);
         size_t nextWrite = (currentWrite + 1) % Size;
-        
+
         if (nextWrite == readIndex.load(std::memory_order_acquire)) {
             return false; // Buffer full
         }
-        
+
         buffer[currentWrite] = item;
         writeIndex.store(nextWrite, std::memory_order_release);
         return true;
@@ -76,11 +75,11 @@ public:
     // Pop element (returns false if buffer is empty)
     bool pop(T& item) {
         size_t currentRead = readIndex.load(std::memory_order_relaxed);
-        
+
         if (currentRead == writeIndex.load(std::memory_order_acquire)) {
             return false; // Buffer empty
         }
-        
+
         item = buffer[currentRead];
         readIndex.store((currentRead + 1) % Size, std::memory_order_release);
         return true;
@@ -88,8 +87,7 @@ public:
 
     // Check if buffer is empty
     bool isEmpty() const {
-        return readIndex.load(std::memory_order_acquire) == 
-               writeIndex.load(std::memory_order_acquire);
+        return readIndex.load(std::memory_order_acquire) == writeIndex.load(std::memory_order_acquire);
     }
 
     // Check if buffer is full
@@ -119,9 +117,7 @@ public:
     }
 
     // Maximum usable capacity (one slot is reserved to disambiguate full/empty)
-    static constexpr size_t capacity() noexcept {
-        return Size - 1;
-    }
+    static constexpr size_t capacity() noexcept { return Size - 1; }
 
 private:
     alignas(64) T buffer[Size];
@@ -134,29 +130,26 @@ private:
 // =============================================================================
 class ThreadPool {
 public:
-    ThreadPool(size_t numThreads = std::thread::hardware_concurrency()) 
-        : stop(false) {
+    ThreadPool(size_t numThreads = std::thread::hardware_concurrency()) : stop(false) {
         for (size_t i = 0; i < numThreads; ++i) {
             workers.emplace_back([this] {
                 while (true) {
                     std::function<void()> task;
                     {
                         std::unique_lock<std::mutex> lock(queueMutex);
-                        condition.wait(lock, [this] { 
-                            return stop || !tasks.empty(); 
-                        });
-                        
+                        condition.wait(lock, [this] { return stop || !tasks.empty(); });
+
                         if (stop && tasks.empty()) {
                             return;
                         }
-                        
+
                         task = std::move(tasks.front());
                         tasks.pop();
                     }
                     task();
                 }
             });
-            
+
 #ifdef _WIN32
             // Set thread priority to HIGHEST to ensure audio processing isn't starved
             // by UI or background tasks. Using extern defs to avoid windows.h pollution.
@@ -180,8 +173,7 @@ public:
     }
 
     // Enqueue a task
-    template<typename F>
-    void enqueue(F&& task) {
+    template <typename F> void enqueue(F&& task) {
         {
             std::unique_lock<std::mutex> lock(queueMutex);
             tasks.emplace(std::forward<F>(task));
@@ -190,9 +182,7 @@ public:
     }
 
     // Get number of worker threads
-    size_t size() const {
-        return workers.size();
-    }
+    size_t size() const { return workers.size(); }
 
 private:
     std::vector<std::thread> workers;
@@ -210,22 +200,18 @@ public:
     explicit Barrier(int count) : counter(count) {}
 
     // Reset barrier for N expected signals
-    void reset(int count) {
-        counter.store(count, std::memory_order_release);
-    }
+    void reset(int count) { counter.store(count, std::memory_order_release); }
 
     // Signal completion of one task
-    void signal() {
-        counter.fetch_sub(1, std::memory_order_acq_rel);
-    }
+    void signal() { counter.fetch_sub(1, std::memory_order_acq_rel); }
 
     // Wait for all tasks to complete (Spin-wait optimized for low-latency audio)
     void wait() {
         while (counter.load(std::memory_order_acquire) > 0) {
             // In a real-time thread, we prefer spinning over sleeping for short waits.
             // But yielding prevents hogging the core if tasks are long.
-            // For audio graph, tasks should be micro-second scale. 
-             std::this_thread::yield(); 
+            // For audio graph, tasks should be micro-second scale.
+            std::this_thread::yield();
         }
     }
 
@@ -240,7 +226,7 @@ private:
 // and waits for them to complete. No mutexes or allocations in dispatch.
 class RealTimeThreadPool {
 public:
-    using TaskFunc = void(*)(void* context, void* taskData);
+    using TaskFunc = void (*)(void* context, void* taskData);
 
     RealTimeThreadPool(size_t numThreads) : m_stop(false), m_taskCount(0), m_activeTasks(0) {
         m_taskCounter.store(0);
@@ -250,7 +236,7 @@ public:
                 MMCSS::setProAudio();
                 workerLoop(static_cast<uint32_t>(i));
             });
-            
+
 #ifdef _WIN32
             // SetThreadPriority(m_workers.back().native_handle(), 2); // Handled by MMCSS wrapper now
 #endif
@@ -261,7 +247,8 @@ public:
         m_stop = true;
         m_signal.notify_all();
         for (auto& w : m_workers) {
-            if (w.joinable()) w.join();
+            if (w.joinable())
+                w.join();
         }
     }
 
@@ -273,7 +260,7 @@ public:
         m_taskCount = count;
         m_taskCounter.store(0, std::memory_order_release);
         m_syncBarrier = syncBarrier;
-        
+
         m_activeTasks.store(count, std::memory_order_release);
         m_signal.notify_all();
     }
@@ -284,20 +271,22 @@ private:
     void workerLoop(uint32_t threadIdx) {
         (void)threadIdx;
         while (!m_stop) {
-            m_signal.wait([this] { return m_stop || m_taskCounter.load(std::memory_order_acquire) < m_taskCount; });
-            
-            if (m_stop) return;
+            m_signal.wait([this] { return m_stop || static_cast<uint32_t>(m_taskCounter.load(std::memory_order_acquire)) < m_taskCount; });
+
+            if (m_stop)
+                return;
 
             while (true) {
                 int idx = m_taskCounter.fetch_add(1, std::memory_order_acq_rel);
-                if (idx >= (int)m_taskCount) break;
+                if (idx >= (int)m_taskCount)
+                    break;
 
                 m_taskFunc(m_context, m_taskData[idx]);
-                
+
                 if (m_syncBarrier) {
                     m_syncBarrier->signal();
                 }
-                
+
                 m_activeTasks.fetch_sub(1, std::memory_order_acq_rel);
             }
         }
@@ -307,8 +296,7 @@ private:
         std::mutex mtx;
         std::condition_variable cv;
         void notify_all() { cv.notify_all(); }
-        template<typename F>
-        void wait(F&& cond) {
+        template <typename F> void wait(F&& cond) {
             std::unique_lock<std::mutex> lock(mtx);
             cv.wait(lock, std::forward<F>(cond));
         }
@@ -316,7 +304,7 @@ private:
 
     std::vector<std::thread> m_workers;
     std::atomic<bool> m_stop;
-    
+
     // Shared task state
     TaskFunc m_taskFunc{nullptr};
     void* m_context{nullptr};
@@ -337,15 +325,13 @@ private:
 class AtomicFlag {
 public:
     AtomicFlag() : flag(false) {}
-    
+
     void set() { flag.store(true, std::memory_order_release); }
     void clear() { flag.store(false, std::memory_order_release); }
     bool isSet() const { return flag.load(std::memory_order_acquire); }
-    
+
     // Test and set (returns previous value)
-    bool testAndSet() {
-        return flag.exchange(true, std::memory_order_acq_rel);
-    }
+    bool testAndSet() { return flag.exchange(true, std::memory_order_acq_rel); }
 
 private:
     std::atomic<bool> flag;
@@ -355,7 +341,7 @@ private:
 class AtomicCounter {
 public:
     AtomicCounter(int initial = 0) : count(initial) {}
-    
+
     int increment() { return count.fetch_add(1, std::memory_order_acq_rel) + 1; }
     int decrement() { return count.fetch_sub(1, std::memory_order_acq_rel) - 1; }
     int get() const { return count.load(std::memory_order_acquire); }
@@ -369,16 +355,14 @@ private:
 class SpinLock {
 public:
     SpinLock() : flag(false) {}
-    
+
     void lock() {
         while (flag.exchange(true, std::memory_order_acquire)) {
             // Spin
         }
     }
-    
-    void unlock() {
-        flag.store(false, std::memory_order_release);
-    }
+
+    void unlock() { flag.store(false, std::memory_order_release); }
 
 private:
     std::atomic<bool> flag;
